Removed background subsampling
I made changes to SubsampleRenderScene and TraceRay that make the
background stay the same regardless of the subsamples.
SubsampleRenderScene checks if either Scene.lastHit or Scene.currentHit
are null. If they are it will render all pixels between the two sampled
pixels fully. This ensures the background is consistent.

I might make this an option in the future because I think there's some
merit to having the background subsampled. It can look neat in some
cases.