Fixed Scene.lastHit / Scene.currentHit assignment
In TraceRay() I wrote:
  `Scene.lastHit = Scene.currentHit
   Scene.currentHit = object`
  (object is the thing TraceRay just hit)
But this was wrong because TraceRay recursively calls itself for
transparent objects and reflections on objects. So each recursive call
would execute the statement above as well.
My original attempt to fix this was simply moving the lines above to the
very end of TraceRay(), after all of the recursive calls.
This solved the problem with Scene.currentHit because even though each
recursive call assigned Scene.currentHit, the first call of TraceRay
ended up being the last one to run that line.
However there was still a bug related to Scene.lastHit, because it was
being assigned to Scene.currentHit on each call before Scene.currentHit
was updated. This was a problem because when TraceRay had any recursive
calls, Scene.currentHit had just been set before returning from that
recursive call, and Scene.lastHit was set to that value before
Scene.currentHit was overwritten.
Effectively, Scene.lastHit was assigned to the object behind a
transparent object or the object hit after a reflection.
What I really wanted was Scene.lastHit to be Scene.currentHit
from the result of TraceRay on the previous pixel. So I just removed the
line Scene.lastHit = Scene.currentHit from TraceRay and into
SubsampleRenderScene. This seems to work the way I intended from the
beginning.